Internal: Interactions - data-handling [ED-21548]
 * extracted validation into a separate handler
